separating cli core and app
